Rename aux into auxiliary
Don't use aux as a directory name since it is a reserved name
on Windows platforms.

This fixes #34.
