+++ typescript/ecommerce-store/services/src/auxiliary/mailgun_client.ts
