+++ typescript/food-ordering/services/src/auxiliary/delivery_provider_api_client.ts
