+++ typescript/ticket-reservation/src/auxiliary/email_client.ts
